// @types/schema.d.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    /** An ISO-8601 encoded date string. */
    Date: any;
    /** An ISO-8601 encoded UTC date string. */
    DateTime: any;
    /** A Git object ID. */
    GitObjectID: any;
    /** Git SSH string */
    GitSSHRemote: any;
    /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
    GitTimestamp: any;
    /** A string containing HTML code. */
    HTML: any;
    /** An ISO-8601 encoded UTC date string with millisecond precision. */
    PreciseDateTime: any;
    /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
    URI: any;
    /** A valid x509 certificate string */
    X509Certificate: any;
};
// ...
export type Actor = {
    /** A URL pointing to the actor's public avatar. */
    avatarUrl: Scalars['URI'];
    /** The username of the actor. */
    login: Scalars['String'];
    /** The HTTP path for this actor. */
    resourcePath: Scalars['URI'];
    /** The HTTP URL for this actor. */
    url: Scalars['URI'];
};
// ...
export type Assignable = {
    /** A list of Users assigned to this object. */
    assignees: UserConnection;
};
// ...
export type Closable = {
    /** `true` if the object is closed (definition of closed may depend on type) */
    closed: Scalars['Boolean'];
    /** Identifies the date and time when the object was closed. */
    closedAt?: Maybe<Scalars['DateTime']>;
};
// ...
export type Comment = {
    /** The actor who authored the comment. */
    author?: Maybe<Actor>;
    /** Author's association with the subject of the comment. */
    authorAssociation: CommentAuthorAssociation;
    /** The body as Markdown. */
    body: Scalars['String'];
    /** The body rendered to HTML. */
    bodyHTML: Scalars['HTML'];
    /** The body rendered to text. */
    bodyText: Scalars['String'];
    /** Identifies the date and time when the object was created. */
    createdAt: Scalars['DateTime'];
    /** Check if this comment was created via an email reply. */
    createdViaEmail: Scalars['Boolean'];
    /** The actor who edited the comment. */
    editor?: Maybe<Actor>;
    id: Scalars['ID'];
    /** Check if this comment was edited and includes an edit with the creation data */
    includesCreatedEdit: Scalars['Boolean'];
    /** The moment the editor made the last edit */
    lastEditedAt?: Maybe<Scalars['DateTime']>;
    /** Identifies when the comment was published at. */
    publishedAt?: Maybe<Scalars['DateTime']>;
    /** Identifies the date and time when the object was last updated. */
    updatedAt: Scalars['DateTime'];
    /** A list of edits to this content. */
    userContentEdits?: Maybe<UserContentEditConnection>;
    /** Did the viewer author this comment. */
    viewerDidAuthor: Scalars['Boolean'];
};


/** Represents a comment. */
export type CommentUserContentEditsArgs = {
    after?: Maybe<Scalars['String']>;
    before?: Maybe<Scalars['String']>;
    first?: Maybe<Scalars['Int']>;
    last?: Maybe<Scalars['Int']>;
};

/** A comment author association with repository. */
export type CommentAuthorAssociation =
/** Author has been invited to collaborate on the repository. */
    | 'COLLABORATOR'
    /** Author has previously committed to the repository. */
    | 'CONTRIBUTOR'
    /** Author has not previously committed to GitHub. */
    | 'FIRST_TIMER'
    /** Author has not previously committed to the repository. */
    | 'FIRST_TIME_CONTRIBUTOR'
    /** Author is a placeholder for an unclaimed user. */
    | 'MANNEQUIN'
    /** Author is a member of the organization that owns the repository. */
    | 'MEMBER'
    /** Author has no association with the repository. */
    | 'NONE'
    /** Author is the owner of the repository. */
    | 'OWNER';
// ...
export type Hovercard = {
    __typename?: 'Hovercard';
    /** Each of the contexts for this hovercard */
    contexts: Array<HovercardContext>;
};
// ...
export type Issue = Assignable & Closable & Comment & Labelable & Lockable & Node & Reactable & RepositoryNode & Subscribable & UniformResourceLocatable & Updatable & UpdatableComment & {
    __typename?: 'Issue';
    /** Reason that the conversation was locked. */
    activeLockReason?: Maybe<LockReason>;
    /** A list of Users assigned to this object. */
    assignees: UserConnection;
    /** The actor who authored the comment. */
    author?: Maybe<Actor>;
    /** Author's association with the subject of the comment. */
    authorAssociation: CommentAuthorAssociation;
    /** Identifies the body of the issue. */
    body: Scalars['String'];
    /** The body rendered to HTML. */
    bodyHTML: Scalars['HTML'];
    /** The http path for this issue body */
    bodyResourcePath: Scalars['URI'];
    /** Identifies the body of the issue rendered to text. */
    bodyText: Scalars['String'];
    /** The http URL for this issue body */
    bodyUrl: Scalars['URI'];
    /** `true` if the object is closed (definition of closed may depend on type) */
    closed: Scalars['Boolean'];
    /** Identifies the date and time when the object was closed. */
    closedAt?: Maybe<Scalars['DateTime']>;
    /** A list of comments associated with the Issue. */
    comments: IssueCommentConnection;
    /** Identifies the date and time when the object was created. */
    createdAt: Scalars['DateTime'];
    /** Check if this comment was created via an email reply. */
    createdViaEmail: Scalars['Boolean'];
    /** Identifies the primary key from the database. */
    databaseId?: Maybe<Scalars['Int']>;
    /** The actor who edited the comment. */
    editor?: Maybe<Actor>;
    /** The hovercard information for this issue */
    hovercard: Hovercard;
    id: Scalars['ID'];
    /** Check if this comment was edited and includes an edit with the creation data */
    includesCreatedEdit: Scalars['Boolean'];
    /** Indicates whether or not this issue is currently pinned to the repository issues list */
    isPinned?: Maybe<Scalars['Boolean']>;
    /** Is this issue read by the viewer */
    isReadByViewer?: Maybe<Scalars['Boolean']>;
    /** A list of labels associated with the object. */
    labels?: Maybe<LabelConnection>;
    /** The moment the editor made the last edit */
    lastEditedAt?: Maybe<Scalars['DateTime']>;
    /** `true` if the object is locked */
    locked: Scalars['Boolean'];
    /** Identifies the milestone associated with the issue. */
    milestone?: Maybe<Milestone>;
    /** Identifies the issue number. */
    number: Scalars['Int'];
    /** A list of Users that are participating in the Issue conversation. */
    participants: UserConnection;
    /** List of project cards associated with this issue. */
    projectCards: ProjectCardConnection;
    /** Identifies when the comment was published at. */
    publishedAt?: Maybe<Scalars['DateTime']>;
    /** A list of reactions grouped by content left on the subject. */
    reactionGroups?: Maybe<Array<ReactionGroup>>;
    /** A list of Reactions left on the Issue. */
    reactions: ReactionConnection;
    /** The repository associated with this node. */
    repository: Repository;
    /** The HTTP path for this issue */
    resourcePath: Scalars['URI'];
    /** Identifies the state of the issue. */
    state: IssueState;
    /**
     * A list of events, comments, commits, etc. associated with the issue.
     * @deprecated `timeline` will be removed Use Issue.timelineItems instead. Removal on 2020-10-01 UTC.
     */
    timeline: IssueTimelineConnection;
    /** A list of events, comments, commits, etc. associated with the issue. */
    timelineItems: IssueTimelineItemsConnection;
    /** Identifies the issue title. */
    title: Scalars['String'];
    /** Identifies the issue title rendered to HTML. */
    titleHTML: Scalars['String'];
    /** Identifies the date and time when the object was last updated. */
    updatedAt: Scalars['DateTime'];
    /** The HTTP URL for this issue */
    url: Scalars['URI'];
    /** A list of edits to this content. */
    userContentEdits?: Maybe<UserContentEditConnection>;
    /** Can user react to this subject */
    viewerCanReact: Scalars['Boolean'];
    /** Check if the viewer is able to change their subscription status for the repository. */
    viewerCanSubscribe: Scalars['Boolean'];
    /** Check if the current viewer can update this object. */
    viewerCanUpdate: Scalars['Boolean'];
    /** Reasons why the current viewer can not update this comment. */
    viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
    /** Did the viewer author this comment. */
    viewerDidAuthor: Scalars['Boolean'];
    /** Identifies if the viewer is watching, not watching, or ignoring the subscribable entity. */
    viewerSubscription?: Maybe<SubscriptionState>;
};
// ...
export type IssueCommentConnection = {
    __typename?: 'IssueCommentConnection';
    /** A list of edges. */
    edges?: Maybe<Array<Maybe<IssueCommentEdge>>>;
    /** A list of nodes. */
    nodes?: Maybe<Array<Maybe<IssueComment>>>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
    /** Identifies the total count of items in the connection. */
    totalCount: Scalars['Int'];
};
// ...
export type IssueState =
/** An issue that has been closed */
    | 'CLOSED'
    /** An issue that is still open */
    | 'OPEN';
// ...
export type Labelable = {
    /** A list of labels associated with the object. */
    labels?: Maybe<LabelConnection>;
};
// ...
export type LockReason =
/** The issue or pull request was locked because the conversation was off-topic. */
    | 'OFF_TOPIC'
    /** The issue or pull request was locked because the conversation was resolved. */
    | 'RESOLVED'
    /** The issue or pull request was locked because the conversation was spam. */
    | 'SPAM'
    /** The issue or pull request was locked because the conversation was too heated. */
    | 'TOO_HEATED';

/** An object that can be locked. */
export type Lockable = {
    /** Reason that the conversation was locked. */
    activeLockReason?: Maybe<LockReason>;
    /** `true` if the object is locked */
    locked: Scalars['Boolean'];
};
// ...
export type Node = {
    /** ID of the object. */
    id: Scalars['ID'];
};
// ...
export type Reactable = {
    /** Identifies the primary key from the database. */
    databaseId?: Maybe<Scalars['Int']>;
    id: Scalars['ID'];
    /** A list of reactions grouped by content left on the subject. */
    reactionGroups?: Maybe<Array<ReactionGroup>>;
    /** A list of Reactions left on the Issue. */
    reactions: ReactionConnection;
    /** Can user react to this subject */
    viewerCanReact: Scalars['Boolean'];
};
// ...
export type RepositoryNode = {
    /** The repository associated with this node. */
    repository: Repository;
};
// ...
export type Subscribable = {
    id: Scalars['ID'];
    /** Check if the viewer is able to change their subscription status for the repository. */
    viewerCanSubscribe: Scalars['Boolean'];
    /** Identifies if the viewer is watching, not watching, or ignoring the subscribable entity. */
    viewerSubscription?: Maybe<SubscriptionState>;
};
// ...
export type SubscriptionState =
/** The User is never notified. */
    | 'IGNORED'
    /** The User is notified of all conversations. */
    | 'SUBSCRIBED'
    /** The User is only notified when participating or @mentioned. */
    | 'UNSUBSCRIBED';
// ...
export type UniformResourceLocatable = {
    /** The HTML path to this resource. */
    resourcePath: Scalars['URI'];
    /** The URL to this resource. */
    url: Scalars['URI'];
};
// ...
export type Updatable = {
    /** Check if the current viewer can update this object. */
    viewerCanUpdate: Scalars['Boolean'];
};

/** Comments that can be updated. */
export type UpdatableComment = {
    /** Reasons why the current viewer can not update this comment. */
    viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
};
// ...
export type UserConnection = {
    __typename?: 'UserConnection';
    /** A list of edges. */
    edges?: Maybe<Array<Maybe<UserEdge>>>;
    /** A list of nodes. */
    nodes?: Maybe<Array<Maybe<User>>>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
    /** Identifies the total count of items in the connection. */
    totalCount: Scalars['Int'];
};